move function to release.py
